backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.3
python-dotenv==1.0.0
python-multipart==0.0.9
PyPDF2==3.0.1
pdfplumber==0.10.3
anthropic==0.18.1
transformers==4.37.2
torch==2.6.0
numpy==1.26.1
pandas==2.1.2
pytest==7.4.3
httpx==0.25.1
aiosqlite==0.19.0
sentencepiece==0.2.0
protobuf==4.25.2